ACL_PyTorch/contrib/cv/video_understanding/FOMM/requirements.txt
scikit-image==0.18.0
certifi
cloudpickle
cycler
dask
decorator
imageio
imageio-ffmpeg
kiwisolver
lockfile
matplotlib
networkx
numpy
onnx
onnx-simplifier==0.4.10
onnxruntime
opencv-python
pandas
pillow
pycparser
pyparsing
pytz
pywavelets
pyyaml
scikit-learn
scipy
six
sympy
toolz
torchaudio
torchvision
tqdm
